requirements.txt
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.2.25
charset-normalizer==3.4.6
dateparser==1.3.0
distro==1.9.0
durationpy==0.10
exceptiongroup==1.3.1
groq==0.37.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.13.0
jsonpatch==1.33
jsonpointer==3.1.0
kubernetes==35.0.0
langchain==1.2.13
langchain-core==1.2.20
langchain-groq==1.1.2
langchain-openai==1.1.11
langchainhub==0.1.21
langgraph==1.1.3
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.12
langsmith==0.7.22
oauthlib==3.3.1
openai==2.29.0
orjson==3.11.7
ormsgpack==1.12.2
packaging==24.2
prometheus-api-client==0.7.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
PyYAML==6.0.3
regex==2026.2.28
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
tenacity==9.1.4
tiktoken==0.12.0
tqdm==4.67.3
types-requests==2.32.4.20260107
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.6.3
uuid_utils==0.14.1
websocket-client==1.9.0
xxhash==3.6.0
zstandard==0.25.0
